### 2023-08-28 #5785 @editorialbot
[PRE REVIEW]: Micro Manager: a Python package for adaptive and flexible two-scale coupling
**Submitting author:** @IshaanDesai (<a href="http://orcid.org/0000-0002-2552-7509">Ishaan Desai</a>)
**Repository:** https://github.com/precice/micro-manager
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v0.3.0
