### 2023-08-28 #5785 @danielskatz
Also, can you explain the relationship between this code and preCICE in a short statement here? I'm not sure I fully understand this, even as stated in the paper.

And, is there any code overlap between this repository and preCICE?
