### 2023-08-28 #5785 @IshaanDesai
:wave: @danielskatz thanks for starting off this process. Last year we (I) wrote a short conference paper introducing the Micro Manager: *Desai, Ishaan, & Bringedal, Carina & Uekermann, Benjamin. A flexible software approach to simulate two-scale coupled problems. ECCOMAS Congress 2022. [10.23967/eccomas.2022.037](https://doi.org/10.23967/eccomas.2022.037)*. The software has matured quite a bit since then, and the intention is to have a citeable journal paper of the software. There have not been any journal papers or other papers about the Micro Manager.

> Also, can you explain the relationship between this code and preCICE in a short statement here?

[preCICE](https://precice.org/) is a library which facilitates black-box partitioned coupling between simulation solvers on the same physical scales, meaning the solvers solve physics on the same physical domain. The Micro Manager is run together with preCICE to allow for coupling between solvers on different scales, and more precisely, allow a user to couple a solver on a coarse scale to a solver on a fine scale.

> And, is there any code overlap between this repository and preCICE?

No, there is no code overlap between the Micro Manager and the [core preCICE library](https://github.com/precice/precice). The Micro Manager is a repository in the [preCICE project](https://github.com/precice) on GitHub, but the code itself is independent.
