### 2023-08-28 #5785 @danielskatz
👋 @IshaanDesai - thanks, this all sounds good.  You might want to see if you think the paper would benefit from a clear statement like you make in the second answer above. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF, if you want. editorialbot commands need to be the first entry in a new comment.

I'll now look for an editor.
