### 2023-08-30 #5785 @philipcardiff
Hi @ jamiebramwell, @mayrmt, @zbeekman, would you like to review this submission?

The submission under consideration is Micro Manager: a Python package for adaptive and flexible two-scale coupling (https://github.com/precice/micro-manager).

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected.

Philip
