### 2023-09-08 fef2403
Remove deepcopy from solverdummy
