+++ examples/cpp-dummy/micro_cpp_dummy.hpp
@@ -17,7 +17,6 @@ public:
     MicroSimulation();
     // solve takes a python dict data, and the timestep dt as inputs, and returns a python dict
     py::dict solve(py::dict macro_write_data, double dt);
-    MicroSimulation __deepcopy__(py::dict memo);
 
     void set_state(py::list state);
     py::list get_state() const;
