### 2023-09-12 #5842 @editorialbot
[REVIEW]: Micro Manager: a Python package for adaptive and flexible two-scale coupling
**Submitting author:** @IshaanDesai (<a href="http://orcid.org/0000-0002-2552-7509">Ishaan Desai</a>)
**Repository:** https://github.com/precice/micro-manager
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v0.3.0
**Archive:** 10.18419/darus-3764
