### 2023-09-20 #5842 @mrogowski
> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

@IshaanDesai: It seems that this is missing. Let me know if I missed it. Something like a simple `CONTRIBUTING.md` file is enough.
