### 2023-09-20 #5842 @IshaanDesai
@mrogowski yes it is missing, thanks for spotting this! I will add a `CONTRIBUTING.md` file.
