### 2023-09-21 cdddd21
Small fixes to JOSS paper draft (#59)
