+++ paper.md
@@ -26,40 +26,40 @@ bibliography: paper.bib
 
 # Summary
 
-The Micro Manager facilitates coupling between simulation softwares which solve problems at different physical scales. Broadly speaking, simulation-based analysis is an effective tool to gain insights in real-world scenarios without incurring the high cost of prototyping and testing.
-Complex simulations are oftentimes broken down into simpler components which are resolved by tailor-made software.
+The Micro Manager facilitates coupling between simulation software packages, which solve problems at different physical scales. Broadly speaking, simulation-based analysis is an effective tool to gain insights in real-world scenarios without incurring the high cost of prototyping and testing.
+Complex simulations are oftentimes broken down into simpler components, which are resolved by tailor-made software.
 Such complex simulations can be of multiphysics nature, meaning different physics are solved in different parts of the domain.
-To do such multiphysics simulations, we can couple different softwares together. preCICE [@preCICE_v2] is an open-source coupling library for partitioned multiphysics simulations. However, sometimes the coupling is not just between different physics, but also different physical scales, which is oftentimes referred to as multiscale coupling. To understand how the scales affect each other, both scales need to be solved simultaneously while being coupled in a bi-directional manner.
+To do such multiphysics simulations, we can couple different software packages together. preCICE [@preCICE_v2] is an open-source coupling library for partitioned multiphysics simulations. However, sometimes the coupling is not just between different physics but also different physical scales, which is frequently referred to as multiscale coupling. To understand how the scales affect each other, both scales need to be solved simultaneously while being coupled in a bi-directional manner.
 The Micro Manager, together with preCICE, is capable of handling a multiphysics and multiscale coupling.
 We refer to the coarse scale as the macro scale, and the fine scale as the micro scale. The name *Micro Manager* is derived from its core functionality of controlling a set of micro-scale simulations and coupling them to one macro-scale simulation via preCICE.
 
 # Statement of need
 
-Two-scale coupled simulations have already been done in several application areas, such as, porous media  [e.g. @Bastidas_two_scale; @Gaerttner_two_scale], computational mechanics [e.g. @Fritzen_adaptivity] and biomechanics [e.g. @Lambers_liver_multiscale].
+Two-scale coupled simulations have already been done in several application areas, such as porous media  [e.g. @Bastidas_two_scale; @Gaerttner_two_scale], computational mechanics [e.g. @Fritzen_adaptivity] and biomechanics [e.g. @Lambers_liver_multiscale].
 For each of these publications, the coupling software is implemented from scratch. Such implementations typically involve communication between the scales, coupling schemes, and other case-specific technical solutions.
 For coupled problems on a single scale, preCICE handles these coupling aspects.
 The Micro Manager is a thin layer on top of preCICE, which enables preCICE to couple problems across two scales.
 Compared to existing multiscale coupling software such as [MUSCLE3](https://github.com/multiscale/muscle3) [@MUSCLE3], [MUI](https://github.com/MxUI/MUI) [@MUI], and [AMUSE](https://github.com/amusecode/amuse) [@Amuse], the Micro Manager is an add-on package to the general coupling library preCICE and not a stand-alone coupling solution.
 
 For single-scale simulations, many widely-used solvers such as OpenFOAM, FEniCS, deal.II and more, are already coupled using preCICE. Additionally, preCICE has a steadily growing community [@preCICE_v2]. Using the Micro Manager, we expose these advantages of using preCICE for multiscale scenarios.
[…]
