### 2023-09-21 #5842 @IshaanDesai
@mrogowski I added a `CONTRIBUTING.md` file.
