### 2023-10-10 #5842 @IshaanDesai
Hi @mbkumar

We have associated the Micro Manager [v0.3.0](https://github.com/precice/micro-manager/releases/tag/v0.3.0) with this paper. If you install the Micro Manager using pip, the corresponding `pyprecice` (v2.5.0.2) will be automatically.

We do not have a dev container to work with out of the box, but if things get too difficult, we can set something up.
