### 2023-10-11 #5842 @mbkumar
I have tested the software and found it to be rough around the edges. It could be improved in a few aspects. For example, instead of using a fixed name for the MicroManager class, it could be made an abstract class. Then users could subclass it and the macro-manager could be supplied the subclass name in its config file. But, the software is useful and useable in its present form.  The manuscript is reasonably good. So, I recommend its publication.
