### 2023-10-11 #5842 @IshaanDesai
Thank you @mbkumar for the testing the Micro Manager, opening issues for things that did not work, and reviewing the manuscript. We really appreciate it!

At some point we had a configurable class name for the `MicroSimulation` name that the Micro Manager demands right now. To keep things simple and avoid user-side mistakes, we fixed this name. I will open an issue to discuss this once again.
