### 2023-10-12 precice/micro-manager#65 @IshaanDesai
Provide a micro simulation abstract class which users can inherit/create a subclass
In https://github.com/openjournals/joss-reviews/issues/5842#issuecomment-1758196669, @mbkumar suggested that the Micro Manager could provide an abstract class for the micro simulation code, which the users could subclass or inherit, and the give the name of their class to the Micro Manager via the JSON config file. This is a cleaner way to ensure that the users stick to the desired API while converting their micro simulations into callable libraries.
