+++ tests/integration/test_unit_cube/micro_dummy.py
@@ -6,10 +6,11 @@ In this script we solve a dummy micro problem to just show the working of the ma
 
 class MicroSimulation:
 
-    def __init__(self):
+    def __init__(self, sim_id):
         """
         Constructor of MicroSimulation class.
         """
+        self._sim_id = sim_id
         self._micro_scalar_data = None
         self._micro_vector_data = None
         self._checkpoint = None
