### 2023-10-13 9e2fc8d
Pass ID to micro simulation class constructors in dummies
