### 2023-10-13 b656da4
Pass the global ID in the state of the C++ micro simulation in the dummy
