### 2023-10-13 086ad8c
Fix C++ solver dummy
