### 2023-10-14 fcd2b29
Clean up adaptivity tests and some renaming (#67)
* Clean up adaptivity tests

* Rename long variables in micro_manager.py
