### 2023-10-15 e987039
Remove function set_global_id from Simulation class, as we should never have to change the global ID of a micro simulation
