### 2023-10-22 #5842 @mrogowski
I am trying to install the Micro Manager, but I am running into issues. I am runnning in an Ubuntu 22.04 container, and installed preCICE and Micro Manager as follows:
```
apt install ./libprecice2_2.5.0_jammy.deb
pip install --user micro-manager-precice
```
Now, when I try to execute some of the tests such as `test_micro_manager.py` or `unit_cube.py`, I get:
```AttributeError: module 'precice' has no attribute 'Interface'``` and ```AttributeError: module 'precice' has no attribute 'Participant'```.

`precice` is installed:
```
>>> precice.get_version_information()
b'2.5.0;no-info [git failed to run];PRECICE_MPICommunication=Y;PRECICE_PETScMapping=Y;PRECICE_PythonActions=Y;PRECICE_ENABLE_C=Y;PRECICE_ENABLE_FORTRAN=Y;CXX=GNU;CXXFLAGS= -O3 -DNDEBUG;LDFAGS='
```

Can you please advise @IshaanDesai?
