### 2023-10-22 #5842 @IshaanDesai
Hi @mrogowski,

When you cloned the [micro-manager](https://github.com/precice/micro-manager), the default `develop` branch is cloned. Please move to the [v0.3.0](https://github.com/precice/micro-manager/releases/tag/v0.3.0) to run the tests, because this version is compatible with preCICE v2.5.0.
