### 2023-10-23 749cb0e
Add --config parameter to the run-micro-manager.py execution command
