### 2023-10-26 #5842 @mrogowski
We still have a small discussion regarding decomposition for multiprocess simulations going in https://github.com/precice/micro-manager/pull/69, but I consider the checklist complete.

I would appreciate it if there was a test to check if the package has been installed correctly included when Micro Manager is installed with pip (something like `python -m micro_manager --test` and `--mpi_test`) since I could not initially run the examples included in the repository due to version mismatches. I also hope we will soon resolve the decomposition issues from https://github.com/precice/micro-manager/pull/69.

Having said that, I am happy to recommend Micro Manager for publication as is.
