### 2023-10-29 f45efbd
README clarification + fix typo (#69)
* Fix typo

* Add --config parameter to the run-micro-manager.py execution command

* Add assertion to check if a rank has no vertices

---------

Co-authored-by: Ishaan Desai <ishaandesai@gmail.com>
