### 2023-10-29 c9d9b94
Add reference to Helena Kschidocks contribution and the DuMux solver
