+++ paper.bib
@@ -167,6 +167,14 @@
 	author = "Fritz Otlinghaus"
 }
 
+@report{Kschidock2023DuMuxMacroMicro,
+	title = "Macro-micro coupling with DuMux and preCICE",
+	year = "2023",
+	school = "University of Stuttgart",
+	doi = "10.18419/opus-13560",
+	author = "Helena Kschidock"
+}
+
 @article{Gaerttner_two_scale,
     author = {Gärttner, Stephan and Frolkovič, Peter and Knabner, Peter and Ray, Nadja},
     title = {Efficiency and Accuracy of Micro-Macro Models for Mineral Dissolution},
@@ -176,4 +184,16 @@
     pages = {e2020WR027585},
     doi = {10.1029/2020WR027585},
     year = {2020}
+}
+
+@article{Koch2021DuMux,
+	title = {DuMux 3 – an open-source simulator for solving flow and transport problems in porous media with a focus on model coupling},
+	journal = {Computers & Mathematics with Applications},
+	volume = {81},
+	pages = {423-443},
+	year = {2021},
+	note = {Development and Application of Open-source Software for Problems with Numerical PDEs},
+	issn = {0898-1221},
+	doi = {10.1016/j.camwa.2020.02.012},
+	author = {Timo Koch and Dennis Gläser and Kilian Weishaupt and Sina Ackermann and Martin Beck and Beatrix Becker and Samuel Burbulla and Holger Class and Edward Coltman and Simon Emmert and Thomas Fetzer and Christoph Grüninger and Katharina Heck and Johannes Hommel and Theresa Kurz and Melanie Lipp and Farid Mohammadi and Samuel Scherrer and Martin Schneider and Gabriele Seitz and Leopold Stadler and Martin Utz and Felix Weinhardt and Bernd Flemisch}
 }
