### 2023-11-04 #5842 @IshaanDesai
Version of the Micro Manager: [v0.3.0](https://github.com/precice/micro-manager/releases/tag/v0.3.0)
DOI: https://doi.org/10.18419/darus-3764
