### 2023-11-07 precice/micro-manager#70 @IshaanDesai
Add check if dependencies are installed correctly when Micro Manager is installed via pip
As suggested in https://github.com/openjournals/joss-reviews/issues/5842#issuecomment-1781849246 there needs to be a check to ensure that the dependencies are installed correctly when the Micro Manager is installed via pip.
