### 2023-11-07 #5842 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
