### 2023-11-07 #5842 @danielskatz
I've suggested a few small changes in https://github.com/precice/micro-manager/pull/71 - please merge this, or let me know what you disagree with, then we can continue towards acceptance and publication
