### 2023-11-07 #5842 @danielskatz
Congratulations to @IshaanDesai (Ishaan Desai) and co-authors!!

JOSS depends on volunteers and couldn't be successful without you
